+++ typescript/sdk/src/core/EvmCoreModule.ts
@@ -23,7 +23,7 @@ import {
   HyperlaneAddresses,
   HyperlaneContractsMap,
 } from '../contracts/types.js';
-import { DeployedCoreAddresses } from '../core/schemas.js';
+import { CoreConfigSchema, DeployedCoreAddresses } from '../core/schemas.js';
 import { CoreConfig, DerivedCoreConfig } from '../core/types.js';
 import { HyperlaneProxyFactoryDeployer } from '../deploy/HyperlaneProxyFactoryDeployer.js';
 import {
@@ -49,7 +49,6 @@ import { EvmCoreReader } from './EvmCoreReader.js';
 import { EvmIcaModule } from './EvmIcaModule.js';
 import { HyperlaneCoreDeployer } from './HyperlaneCoreDeployer.js';
 import { CoreFactories } from './contracts.js';
-import { CoreConfigSchema } from './schemas.js';
 
 export class EvmCoreModule extends HyperlaneModule<
   ProtocolType.Ethereum,