+++ typescript/sdk/src/token/adapters/EvmTokenAdapter.ts
@@ -261,9 +261,7 @@ export class EvmHypCollateralAdapter
     return this.wrappedTokenAddress!;
   }
 
-  protected async getWrappedTokenAdapter(): Promise<
-    ITokenAdapter<PopulatedTransaction>
-  > {
+  protected async getWrappedTokenAdapter(): Promise<EvmTokenAdapter> {
     return new EvmTokenAdapter(this.chainName, this.multiProvider, {
       token: await this.getWrappedTokenAddress(),
     });
@@ -301,6 +299,21 @@ export class EvmHypCollateralAdapter
     return this.getWrappedTokenAdapter().then((t) =>
       t.populateTransferTx(params),
     );
+  }
+}
+
+export class EvmHypCollateralFiatAdapter
+  extends EvmHypCollateralAdapter
+  implements IHypTokenAdapter<PopulatedTransaction>
+{
+  /**
+   * Note this may be inaccurate, as this returns the total supply
+   * of the fiat token, which may be used by other bridges.
+   * However this is the best we can do with a simple view call.
+   */
+  override async getBridgedSupply(): Promise<bigint> {
+    const wrapped = await this.getWrappedTokenAdapter();
+    return wrapped.getTotalSupply();
   }
 }